refactor(transcription): split tasks/transcription/core.py into 12 modules
Pure mechanical extraction for #284 A3.5 — no behaviour change. All 40
top-level defs move verbatim (verified AST-identical against HEAD) out of the
2901-line module:

  context          TranscriptionContext + the failure/validation handlers
  user_settings    per-user language + tuning prefs
  pipelines        local WhisperX / Engine fast path / transcribe-only stage
  cloud_asr        cloud provider pipeline + parallel cloud diarization
  embeddings       native + traditional speaker embeddings, v4 staging
  finalize         critical path: segments -> speakers -> DB
  background       the post-GPU daemon-thread work
  downstream       summarization / topics / clustering dispatch
  legacy_task      transcription.process_file (the pre-chain monolith)
  diarize_task     transcription.diarize_gpu (multi-GPU split stage 2b)
  cpu_task         transcription.cpu_transcribe

core.py keeps transcription.gpu_transcribe and stays the import surface for
dispatch.py, preprocess.py, postprocess.py, rediarize_task.py and celery_app's
include= list; the symbols those modules import are re-exported.

The registered Celery task set is byte-identical to master (30 tasks), which
is what proves the split did not drop a task registration.
